kalosm

Module language

Kalosm Language

Language processing utilities for the Kalosm framework.

The language part of Kalosm has a few core parts:

Models: Text generation and embedding models

Context: Document collection, format support, search and chunking

Integrations: SurrealDB, Serper, and other integrations

Text Generation Models

[Model] and [ModelExt] are the core traits for text generation models. Any model that implements these traits can be used with Kalosm.

The simplest way to use a model is to create a [Model] and call [ModelExt::stream_text] to stream text from it:

use kalosm::language::*;
#[tokio::main]
async fn main() {
    let mut llm = Llama::new().await.unwrap();
    let prompt = "The following is a 300 word essay about why the capital of France is Paris:";
    print!("{prompt}");
    // Any model that implements the [`TextCompletionModel`] trait can be used to stream text
    let mut stream = llm
        .complete(prompt);
    // You can then use the stream however you need. to_std_out will print the text to the console as it is generated
    stream.to_std_out().await.unwrap();
}

Tasks

You can define a Task with a description then run it with an input. The task will cache the description to repeated calls faster. Tasks work with chat models.

// Create a new model
let model = Llama::new_chat().await.unwrap();
// Create a new task that summarizes text
let task = model.task("You take a long description and summarize it into a single short sentence");
let mut output = task("You can define a Task with a description then run it with an input. The task will cache the description to repeated calls faster. Tasks work with chat models.");
// Then stream the output to the console
output.to_std_out().await.unwrap();

Structured Generation

Structured generation gives you more control over the output of the text generation. You can derive a parser for your data to easily get structured data out of an LLM:

use kalosm::language::*;
#[derive(Parse, Clone)]
struct Pet {
    name: String,
    age: u32,
    description: String,
}

Then you can generate text that works with the parser in a Task:

#[tokio::main]
async fn main() {
    // First create a model. Chat models tend to work best with structured generation
    let model = Llama::new_chat().await.unwrap();
    // Then create a parser for your data. Any type that implements the `Parse` trait has the `new_parser` method
    let parser = Pet::new_parser();
    // Then create a task with the parser as constraints
    let task = model.task("You generate realistic JSON placeholders")
        .with_constraints(Arc::new(parser));
    // Finally, run the task
    let pet: Pet = task("Generate a pet in the form {\"name\": \"Pet name\", \"age\": 0, \"description\": \"Pet description\"}").await.unwrap();
    println!("{pet:?}");
}

Embedding Models

Embedder and EmbedderExt are the core traits for text embedding models. Any model that implements these traits can be used with Kalosm.

The simplest way to use an embedding model is to create an Embedder and call EmbedderExt::embed to create an Embedding of the text which represents the meaning of the text in a numerical format:

use kalosm::language::*;
#[tokio::main]
async fn main() {
    // First create a model. Bert::new() is a good default embedding model for general tasks
    let model = Bert::new().await.unwrap();
    // Then embed some text into the vector space
    let embedding = model.embed("Kalosm is a library for building AI applications").await.unwrap();
    // And some more text
    let embedding = model.embed(prompt_input("Text: ").unwrap()).await.unwrap();
    // You can compare the cosine similarity of the two embeddings to see how similar they are
    println!("cosine similarity: {}", embedding.cosine_similarity(&embedding));
}

Context

Gathering context is a key part of building LLM applications. Providing the right context to the model makes the output more relevant and useful. It can also help to prevent hallucinations.

Kalosm provides tools to generate gather, and process context from a variety of sources.

Gathering context

Kalosm provides utilities for collecting context from a variety of sources:

Local files (.txt, .md, .html, .docx, .pdf)

RSS feeds

Websites

Search engines

Microphone input and audio input through whisper transcriptions

Each of these sources implement either IntoDocument or IntoDocuments to convert the data into a Document with the contents and metadata about the document.

use kalosm::language::*;
use std::convert::TryFrom;
use std::path::PathBuf;

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    // Try to extract an article from a URL
    let page = Url::parse("https://www.nytimes.com/live/2023/09/21/world/zelensky-russia-ukraine-news")?;
    let document = page.into_document().await?;
    println!("Title: {}", document.title());
    println!("Body: {}", document.body());

    Ok(())
}

Chunking context

After you have gathered context, it is often useful to chunk it into smaller pieces for search. Kalosm provides utilities for chunking context into documents, sentences, paragraphs, or semantic chunks. Kalosm will embed each chunk as it splits the document into smaller pieces. One of the most powerful chunker is the semantic chunker, which lets you chunk documents into semantically similar chunks without explicitly setting the size of the chunks:

use kalosm::language::*;

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    // First, create an embedding model for semantic chunking
    let model = Bert::new().await?;
    // Then create a document folder with some documents
    let documents = DocumentFolder::new("./documents")?.into_documents().await?;
    // Then chunk the documents into sentences
    let chunked = SemanticChunker::new().chunk_batch(&documents, &model).await?;
    println!("{:?}", chunked);
    Ok(())
}

Embedding-powered search

After you have chunked your context, you can use the embeddings for search or retrieval augmented generation. Embedding-based search lets you find documents that are semantically similar to a specific word or phrase even if no words are an exact match:

use kalosm::language::*;
use surrealdb::{engine::local::SurrealKv, Surreal};

#[tokio::main]
async fn main() {
    // Create database connection
    let db = Surreal::new::<SurrealKv>(std::env::temp_dir().join("temp.db")).await.unwrap();

    // Select a specific namespace / database
    db.use_ns("search").use_db("documents").await.unwrap();

    // Create a table in the surreal database to store the embeddings
    let document_table = db
        .document_table_builder("documents")
        .build::<Document>()
        .await
        .unwrap();

    // Add documents to the database
    document_table.add_context(DocumentFolder::new("./documents").unwrap()).await.unwrap();

    loop {
        // Get the user's question
        let user_question = prompt_input("Query: ").unwrap();

        let nearest_5 = document_table
            .search(&user_question)
            .with_results(5)
            .await
            .unwrap();

        println!("{:?}", nearest_5);
    }
}

Re-exports

pub use kalosm_language::kalosm_language_model::ChatModel as _;

pub use kalosm_language::kalosm_language_model::ChatModelExt as _;

pub use kalosm_language::kalosm_language_model::ChatSession as _;

pub use kalosm_language::kalosm_language_model::CreateChatSession as _;

pub use kalosm_language::kalosm_language_model::CreateDefaultChatConstraintsForType as _;

pub use kalosm_language::kalosm_language_model::CreateDefaultCompletionConstraintsForType as _;

pub use kalosm_language::kalosm_language_model::CreateTextCompletionSession as _;

pub use kalosm_language::kalosm_language_model::Embedder as _;

pub use kalosm_language::kalosm_language_model::EmbedderCacheExt as _;

pub use kalosm_language::kalosm_language_model::EmbedderExt as _;

pub use kalosm_language::kalosm_language_model::IntoChatMessage as _;

pub use kalosm_language::kalosm_language_model::IntoEmbedding as _;

pub use kalosm_language::kalosm_language_model::ModelBuilder as _;

pub use kalosm_language::kalosm_language_model::ModelConstraints as _;

pub use kalosm_language::kalosm_language_model::StreamExt as _;

pub use kalosm_language::kalosm_language_model::StructuredChatModel as _;

pub use kalosm_language::kalosm_language_model::StructuredTextCompletionModel as _;

pub use kalosm_language::kalosm_language_model::TextCompletionModel as _;

pub use kalosm_language::kalosm_language_model::TextCompletionModelExt as _;

pub use kalosm_language::kalosm_language_model::TextCompletionSession as _;

pub use kalosm_language::kalosm_sample;

Modules

kalosm_sample

kalosm-sample

Macros

bail

Bail out with the given error.

Structs

AnyOfSchema

A schema that matches any of the composite schemas

ArcParser

A boxed parser.

ArraySchema

A schema for an array

Attributes

Bert

A bert embedding model. The main interface for this model is EmbedderExt.

BertBuilder

A builder for a Bert model

BertSource

A the source of a crate::Bert model

BooleanSchema

A schema for a boolean

BoxedChatConstraintsForType

A constraints for CreateDefaultChatConstraintsForType that work with boxed StructuredChatModels.

BoxedChatModel

A boxed ChatModel.

BoxedChatSession

A boxed ChatSession.

BoxedCompletionConstraintsForType

A constraints for CreateDefaultCompletionConstraintsForType that work with boxed TextCompletionModels.

BoxedStructuredChatModel

A boxed StructuredChatModel.

BoxedStructuredTextCompletionModel

A boxed BoxedStructuredTextCompletionModel.

BoxedTextCompletionModel

A boxed TextCompletionModel.

BoxedTextCompletionSession

A boxed TextCompletionSession.

CachedEmbeddingModel

Embedding models can be expensive to run. This struct wraps an embedding model with a cache that stores embeddings that have been computed before.

ChannelTextStream

A stream of text from a tokio channel.

Chat

Chat is a chat interface that builds on top of crate::ChatModel and crate::StructuredChatModel. It makes it easy to create a chat session with streaming responses, and constraints. Let’s start with a simple chat application:

ChatMessage

A single item in the chat history.

ChatResponseBuilder

A builder for a chat response. This is returned by Chat::add_message and can be modified until you start awaiting the response.

ChoiceParser

A parser for a choice of two parsers.

ChoiceParserState

State of a choice parser.

Chunk

A document snippet that can be used to display a snippet of a document.

ClosureMarker

A marker type that allows kalosm to specialize the IntoVectorDbSearchFilter trait for closures.

ConstSchema

A schema for a constant

DefaultSentenceChunker

The default sentence chunker. Unlike SentenceChunker, this is Send + Sync

Document

A document is a piece of text with a title.

DocumentFolder

A folder full of documents.

DocumentFolderNotDirectoryError

The path to a document folder was not a directory

DocxDocument

A docx document that can be read from the file system.

DynEmbedder

A trait object for an embedder.

Embedding

Embeddings

EmbeddingId

A unique identifier for an embedding. If you delete an embedding, the id will be recycled.

EmbeddingInput

The input to an embedding model. This includes the text to be embedded and the type of embedding to output.

EmptyNumber

An error that can occur while parsing a float literal when trying to parse a number with no characters.

EnumSchema

A schema for an enum

FileLoadingProgress

The progress of a file download

FloatParser

A parser for a float.

FloatParserState

The state of an integer parser.

GenerationParameters

Parameters to use when generating text.

Html

An HTML tree.

HtmlDocument

An html document that can be read from the file system.

HtmlSimplifier

Simplifies HTML by removing attributes and elements that does not affect the content

Hypothetical

Generates questions for a document.

HypotheticalBuilder

A builder for a hypothetical chunker.

I8Parser

A parser for i8.

I16Parser

A parser for i16.

I32Parser

A parser for i32.

I64Parser

A parser for i64.

IfThenSchema

A schema for an conditional schema

IndexParser

A parser that parses a sequence of parsers and returns the index of the first parser that succeeds

IndexParserState

The state of the IndexParser parser

IntegerParser

A parser for an integer.

IntegerParserState

The state of an integer parser.

IntegerSchema

A schema for an integer

InvalidDecimalLocation

An error that can occur while parsing a float literal when the number contains a decimal point in the wrong place.

InvalidSignLocation

An error that can occur while parsing a float literal when the number contains a sign in the wrong place.

IteratorMarker

A marker type that allows kalosm to specialize the IntoVectorDbSearchFilter trait for iterators.

JsonObjectSchema

A schema for an object

JsonPropertySchema

A schema for a property of an object

KnowledgeGraph

LazyParser

A parser that is lazily initialized.

LeadingZeroError

An error that can occur while parsing a float literal when the number starts with a leading zero.

LiteralMismatchError

The error type for a literal parser.

LiteralParser

A parser for a literal.

LiteralParserOffset

The state of a literal parser.

Llama

A quantized Llama language model with support for streaming generation.

LlamaBuilder

A builder with configuration for a Llama model.

LlamaChatSession

A Llama chat session.

LlamaSession

A Llama session with cached state for the current fed prompt

LlamaSource

A source for the Llama model.

MapOutputParser

A parser that maps the output of another parser.

MdDocument

A markdown document that can be read from the file system.

NoConstraints

A marker type that indicates that no constraints were supplied.

NumberSchema

A schema for a number (floating point or integer)

OneLine

One line of text with some non-whitespace characters

OneLineError

An error that can occur when parsing a OneLine

OneLineState

The state of the OneLine parser

OneOfSchema

A schema that matches one of the composite schemas

Organic

OutOfRangeError

An error that can occur while parsing a float literal when the number is out of range.

ParagraphStream

A stream that output paragraphs of text at a time.

ParserError

An error that occurred while parsing.

PdfDocument

A pdf document that can be read from the file system.

PeopleAlsoAsk

RegexParser

A parser that uses a regex pattern to parse input.

RegexParserState

The state of a regex parser.

RelatedSearches

RepeatParser

A parser for a repeat of two parsers.

RepeatParserState

State of a repeat parser.

RssFeed

A RSS feed that can be used to add documents to a search index.

SearchQuery

A search query that can be used to search for documents on the web.

SearchResult

SegmentedStream

A stream that output segments of text at a time.

SemanticChunker

A chunker that tries to create chunks of wroughly the same size while grouping together chunks with a similar meaning.

Sentence

A single word.

SentenceChunker

A Chunker that splits a string into sentences with a given SRX rules.

SentenceParser

A parser for a word.

SentenceStream

A stream that output sentences of text at a time.

SeparatedParser

A parser for a repeat of two parsers.

SeparatedParserState

State of a repeat parser.

SequenceParser

A parser for a sequence of two parsers.

Sitelinks

StopOn

A parser that parses until a literal is found.

StopOnOffset

The state of a stop on literal parser.

StopOnParseError

An error that can occur while parsing a string literal.

StringParseError

An error that can occur while parsing a string literal.

StringParser

A parser for an ascii string.

StringParserState

The state of a literal parser.

StringSchema

A schema for a string

Summarizer

Generates summaries for a document.

Task

A task session lets you efficiently run a task with a model. The task session will reuse the model’s cache to avoid re-feeding the task prompt repeatedly.

TextCompletionBuilder

A builder for a text completion response. This is returned by TextCompletionModelExt::complete and can be modified with TextCompletionBuilder::with_sampler and TextCompletionBuilder::with_constraints until you start awaiting the response.

TextDocument

A text document that can be read from the file system.

ThenLazy

A parser that is initialized lazily based on the state of the previous parser.

U8Parser

A parser for u8.

U16Parser

A parser for u16.

U32Parser

A parser for u32.

U64Parser

A parser for u64.

Url

A parsed URL record.

VectorDB

A vector database that can be used to store embeddings and search for similar embeddings.

VectorDBSearchBuilder

A builder for searching for embeddings in a vector database.

VectorDBSearchResult

A resulting point from a search.

WithInitialState

A parser with lazy initial state

Word

A single word.

WordParser

A parser for a word.

WordStream

A stream that output words of text at a time.

Enums

ChunkStrategy

A strategy for chunking a document into smaller pieces.

Either

A value that can be one of two types.

EmbeddingVariant

The type of embedding the model should output. For models that output different embeddings for queries and documents, this

ExtractDocumentError

An error that can occur when extracting a document from a URL.

FileSource

A source for a file, either from Hugging Face or a local path

FsDocument

A document that can be read from the file system.

FsDocumentError

An error that can occur when reading a document from the file system.

HypotheticalChunkerError

An error that can occur when chunking a document with [HypotheticalChunker].

Lang

Represents a language following ISO 639-3 standard.

MessageType

The type of a chat message

ModelLoadingProgress

The progress starting a model

OwnedParseResult

A parser for a choice between two parsers.

ParseStatus

The state of a parser.

RssFeedError

An error that can occur when interacting with an RSS feed.

SchemaLiteral

A literal value in a schema

SchemaType

The type of a schema

SeparatedItemState

The state of the item in the separated parser.

SequenceParserState

State of a sequence parser.

StructureParser

A validator for a string

StructureParserState

The state of a structure parser.

SummaryChunkerError

An error that can occur when chunking a document with [SummaryChunker].

TextFileDecodeError

An error that can occur when decoding a text file.

ThenLazyParserState

State of a then lazy parser.

VectorDbError

An error that can occur when adding or searching for an embedding to the vector database.

Traits

ChatModel

A trait for unstructured chat models. This trait is required for any chat models even if they do not support structured generation. While this trait is implemented for all chat models, most methods to use models that implement this trait are implemented in the ChatModelExt trait.

ChatModelExt

An extension trait for chat models with helpers for handling chat sessions. This trait is implemented automatically for all crate::ChatModels.

ChatSession

Chat Session

Chunker

A strategy for chunking a document into smaller pieces.

CreateChatSession

A trait for creating a chat session. While it the core trait every chat session implementation implements, most methods to use models that implement this trait are implemented in the ChatModelExt trait.

CreateDefaultChatConstraintsForType

A trait that defines the default constraints for a type with this chat model.

CreateDefaultCompletionConstraintsForType

A trait that defines the default constraints for a type with this model.

CreateParserState

A trait for a parser with a default state.

CreateTextCompletionSession

A trait for creating a text completion session for a model. While it the core trait every text completion model must implement, most methods to use models that implement this trait are implemented in the TextCompletionModelExt trait.

Embedder

A model that can be used to embed text. This trait is generic over the vector space that the model uses to help keep track of what embeddings came from which model.

EmbedderCacheExt

An extension trait for Embedder that allows for caching embeddings.

EmbedderExt

An extension trait for Embedder with helper methods for iterators, and types that can be converted into a string.

IntoChatMessage

A trait for types that can be converted into a chat message.

IntoDocument

A trait for types that can be converted into a document.

IntoDocuments

A document that can be added to a search index.

IntoEmbedding

Convert a type into an embedding with an embedding model.

IntoVectorDbSearchFilter

A trait for anything that can be used to filter the results of a vector search.

ModelBuilder

A builder that can create a model asynchronously.

ModelConstraints

A type that can constrain the output of a model into a specific output type.

Parse

Data that can be parsed incrementally.

Parser

An incremental parser for a structured input.

ParserExt

An extension trait for parsers.

Pattern

A pattern that matches a character.

Schema

A description of the format of a type

SendCreateParserState

An auto trait for a Send parser with a default state.

StreamExt

An extension trait for Streams that provides a variety of convenient combinator functions.

StructuredChatModel

A trait for unstructured chat models that support structured generation. While this trait is implemented for all structured chat models, most methods to use models that implement this trait are implemented in the ChatModelExt trait.

StructuredTextCompletionModel

A trait for text completion models that support structured generation. While this trait is implemented for all structured text completion models, most methods to use models that implement this trait are implemented in the TextCompletionModelExt trait.

TextCompletionModel

A trait for unstructured text completion models. This trait is required for any text completion models even if they do not support structured generation. While this trait is implemented for all text completion models, most methods to use models that implement this trait are implemented in the TextCompletionModelExt trait.

TextCompletionModelExt

Text Completion Models

TextCompletionSession

Text Completion Session

TextStream

A stream of text. This is automatically implemented for all streams of something that acts like a string (String, &str).

Functions

accelerated_device_if_available

Create a candle device that uses any available accelerator.

prompt_input

A simple helper function for prompting the user for input.

search

Type Aliases

Candidates

A set of candidates for a vector search.

ParseResult

A result type for parsers.

Derive Macros

Parse

Derive a default JSON parser for a unit value, struct or enum.

Schema