kalosm::language

Trait IntoDocuments

pub trait IntoDocuments {
    type Error: Send + Sync + 'static;

    // Required method
    fn into_documents(
        self,
    ) -> impl Future<Output = Result<Vec<Document>, Self::Error>> + Send;
}

A document that can be added to a search index.

Required Associated Types

type Error: Send + Sync + 'static

The error type that can occur when converting the document into Documents.

Required Methods

fn into_documents( self, ) -> impl Future<Output = Result<Vec<Document>, Self::Error>> + Send

Convert the document into a Document

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.