kalosm

Trait Metric

pub trait Metric<T> {
    const RANGE: RangeInclusive<f64> = _;

    // Required method
    fn distance(
        &mut self,
        first: &T,
        other: &T,
    ) -> impl Future<Output = f64> + Send;
}

A metric is a way to compare two pieces of data. It is used to evaluate the performance of a model.

Provided Associated Constants

const RANGE: RangeInclusive<f64> = _

The range of values that this metric can return.

Required Methods

fn distance(&mut self, first: &T, other: &T) -> impl Future<Output = f64> + Send

Compute the distance between this piece of data and another piece of data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.